+++ app/graphql/types/DistrictType.js
@@ -1,0 +1,11 @@
+const { GraphQLObjectType, GraphQLID, GraphQLString } = require('graphql');
+
+const DistrictType = new GraphQLObjectType({
+  name: 'District',
+  fields: () => ({
+    distrito_id: { type: GraphQLID },
+    distrito_nombre: { type: GraphQLString },
+  }),
+});
+
+module.exports = DistrictType;