+++ CHANGELOG.md
@@ -20,9 +20,12 @@ All notable changes to this project are documented in this file.
 
 - More endpoints are added in swagger.
 - Add endpoint to get department by ID.
-- Add endpoint to get city by ID.
 - Add endpoint to get district by ID.
+- Add endpoint to get city by ID.
 - Add endpoint to get neighborhood by ID.
+- Add GraphQL for the department query.
+- Add GraphQL for the district query.
+- Add GraphQL for the city query.
 
 ### Refactored